pyproject.toml
[project]
name = "apex-inference-service"
version = "0.1.0"
description = "Production-grade ML inference service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "mlflow>=2.10.0",
    "loguru>=0.7.2",
    "python-multipart>=0.0.9"
]

# Updated dependency group syntax to fix warning
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "ruff>=0.2.1",
    "pytest-asyncio>=0.23.5"
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
# Changed from "src" to "." to allow 'import src.app...'
pythonpath = [
    "."
]